Repo setup plus the connector client
A lot of the repo setup (e.g. ci, testing, formatting and linting) was
borrowed from tf-encrypted and what I assume are current best practices
for developing in python.

Some things of note:

- Some preliminary support for type annotations has been added to the
repo but I think using mypy is still a significant engineering challenge
when trying to integrate with third-party packages
- The protobuf file for the data connector was copied from the main cape
repo. In the future we'll need a way to easily share this protobuf
amongst all of our repos. It resides in connector/proto so that the
python imports will work correctly.