Add basic transformations support to python
High level cape function `apply_policies` takes a list of policies, an
entity (i.e. database table name) and a dataframe to transform. It
currently assumes that the schema of the incoming dataframe matches the
schema that the policy/transforms are referencing. It is also missing
transformations "where" support which should be easy to add in
subsequent PRs. There are things we
can do in the future to make this better such as pulling down the
related schema and running some validation (check column types and
names) before the transforms.